refactor: simplify mobile home feed
